Rename component to `RemoteSourceForm`
